check for ref type Template::Document explicitly
previously we only tested for this implicitly. if a template didn't match an
ARRAY ref and was a ref it must be an Template::Document. Since this seems to
be the most common ref type we test for it first as to short circuit. This
should also prevent bugs like rt #13225 from blowing up our code in the future

Signed-off-by: Caleb Cushing <xenoterracide@gmail.com>